move CellarConfig into its own file
